backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-multipart==0.0.9
celery==5.3.6
redis==5.0.1
httpx==0.27.0
python-dotenv==1.0.1
pillow==10.2.0
requests==2.31.0
rasterio==1.3.9
gdal==3.8.1
numpy==1.26.3
zai-sdk
sqlalchemy==2.0.23
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.1.2
alembic==1.12.1
psycopg2-binary==2.9.9
